+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## **Unreleased**
+**Hearthstone**:
+- Added unknown cards shuffled by Deathwing Worldbreaker.
+
 ## **Release v1.51.4 - 2026-04-02**
 **Updated for Hearthstone 35.0.3**